databuilder/requirements.txt
amundsen-databuilder==3.3.2
neo4j-driver==1.7.2
mock
tabulate
pyaml
pyhive
pyinstaller
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
snowflake-connector-python
snowflake-sqlalchemy
SQLAlchemy>=1.3
